Separate toss client from main logic
